# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRunBody": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "examples": [
              "atencion@prod"
            ],
            "type": "string"
          },
          "input": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "lang": {
            "anyOf": [
              {
                "pattern": "^[a-z]{2}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lang"
          },
          "subject": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SubjectRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "agent"
        ],
        "title": "CreateRunBody",
        "type": "object"
      },
      "JsonValue": {
        "anyOf": [
          {
            "type": "boolean"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "type": "string"
          },
          {
            "type": "string"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
      "ProblemBody": {
        "additionalProperties": false,
        "description": "`application/problem+json` (M9 §3.5).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "trace_id"
        ],
        "title": "ProblemBody",
        "type": "object"
      },
      "ResolutionBody": {
        "additionalProperties": false,
        "properties": {
          "handoff_quality": {
            "enum": [
              "useful",
              "incomplete",
              "unnecessary"
            ],
            "title": "Handoff Quality",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "resolution_code": {
            "title": "Resolution Code",
            "type": "string"
          }
        },
        "required": [
          "resolution_code",
          "handoff_quality"
        ],
        "title": "ResolutionBody",
        "type": "object"
      },
      "SubjectRef": {
        "additionalProperties": false,
        "description": "Referencia al sujeto de un run (tipo y referencia opaca) (M0 §2.3).",
        "properties": {
          "kind": {
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "ref": {
            "minLength": 1,
            "title": "Ref",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "ref"
        ],
        "title": "SubjectRef",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "agent-core",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/handoffs/{handoff_ref}": {
      "get": {
        "operationId": "get_handoff",
        "parameters": [
          {
            "in": "path",
            "name": "handoff_ref",
            "required": true,
            "schema": {
              "title": "Handoff Ref",
              "type": "string"
            }
          },
          {
            "description": "Principal firmado (`IdentityVerifier`).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Principal firmado (`IdentityVerifier`).",
              "title": "Authorization"
            }
          },
          {
            "description": "Delegación firmada (solo asesores).",
            "in": "header",
            "name": "X-On-Behalf-Of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Delegación firmada (solo asesores).",
              "title": "X-On-Behalf-Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "credentials_invalid | principal_expired"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "subject_forbidden | agent_forbidden | version_pin_forbidden | delegation_expired | delegation_mismatch | principal_mismatch"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "not_found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "turn_in_progress | handoff_already_resolved | idempotency_conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "run_closed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "rate_limited | cost_budget_exceeded"
          }
        },
        "summary": "Paquete de traspaso"
      }
    },
    "/v1/handoffs/{handoff_ref}/resolution": {
      "post": {
        "operationId": "post_resolution",
        "parameters": [
          {
            "in": "path",
            "name": "handoff_ref",
            "required": true,
            "schema": {
              "title": "Handoff Ref",
              "type": "string"
            }
          },
          {
            "description": "Principal firmado (`IdentityVerifier`).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Principal firmado (`IdentityVerifier`).",
              "title": "Authorization"
            }
          },
          {
            "description": "Delegación firmada (solo asesores).",
            "in": "header",
            "name": "X-On-Behalf-Of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Delegación firmada (solo asesores).",
              "title": "X-On-Behalf-Of"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolutionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "credentials_invalid | principal_expired"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "subject_forbidden | agent_forbidden | version_pin_forbidden | delegation_expired | delegation_mismatch | principal_mismatch"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "not_found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "turn_in_progress | handoff_already_resolved | idempotency_conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "run_closed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "rate_limited | cost_budget_exceeded"
          }
        },
        "summary": "Resuelve un traspaso"
      }
    },
    "/v1/runs": {
      "post": {
        "operationId": "create_run",
        "parameters": [
          {
            "description": "Principal firmado (`IdentityVerifier`).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Principal firmado (`IdentityVerifier`).",
              "title": "Authorization"
            }
          },
          {
            "description": "Delegación firmada (solo asesores).",
            "in": "header",
            "name": "X-On-Behalf-Of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Delegación firmada (solo asesores).",
              "title": "X-On-Behalf-Of"
            }
          },
          {
            "description": "Obligatoria; identifica el intento.",
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Obligatoria; identifica el intento.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "credentials_invalid | principal_expired"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "subject_forbidden | agent_forbidden | version_pin_forbidden | delegation_expired | delegation_mismatch | principal_mismatch"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "not_found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "turn_in_progress | handoff_already_resolved | idempotency_conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "run_closed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "rate_limited | cost_budget_exceeded"
          }
        },
        "summary": "Crea un run"
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Principal firmado (`IdentityVerifier`).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Principal firmado (`IdentityVerifier`).",
              "title": "Authorization"
            }
          },
          {
            "description": "Delegación firmada (solo asesores).",
            "in": "header",
            "name": "X-On-Behalf-Of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Delegación firmada (solo asesores).",
              "title": "X-On-Behalf-Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "credentials_invalid | principal_expired"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "subject_forbidden | agent_forbidden | version_pin_forbidden | delegation_expired | delegation_mismatch | principal_mismatch"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "not_found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "turn_in_progress | handoff_already_resolved | idempotency_conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "run_closed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "rate_limited | cost_budget_exceeded"
          }
        },
        "summary": "Estado resumido de un run"
      }
    },
    "/v1/runs/{run_id}/transcript": {
      "get": {
        "operationId": "get_transcript",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Principal firmado (`IdentityVerifier`).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Principal firmado (`IdentityVerifier`).",
              "title": "Authorization"
            }
          },
          {
            "description": "Delegación firmada (solo asesores).",
            "in": "header",
            "name": "X-On-Behalf-Of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Delegación firmada (solo asesores).",
              "title": "X-On-Behalf-Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "credentials_invalid | principal_expired"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "subject_forbidden | agent_forbidden | version_pin_forbidden | delegation_expired | delegation_mismatch | principal_mismatch"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "not_found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "turn_in_progress | handoff_already_resolved | idempotency_conflict"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "run_closed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "invalid_request"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemBody"
                }
              }
            },
            "description": "rate_limited | cost_budget_exceeded"
          }
        },
        "summary": "Transcript renderizado"
      }
    },
...
  },
...
}